Move transfer observation to TransferCenter
