+++ Core/Barcelona/MediaUploader.swift
@@ -13,8 +13,6 @@ import Logging
 public enum MediaUploadError: CustomNSError, LocalizedError {
     /// Starting the transfer with `IMFileTransferCenter` failed.
     case transferCreationFailed
-    /// Observing the status of an `IMFileTransfer` using notifications failed.
-    case tranferObservationFailed
     /// The underlying `IMFileTransfer` had an error.
     case transferFailed(code: Int64, description: String, isRecoverable: Bool)
     /// Timed out waiting for the transfer to finish.
@@ -24,8 +22,6 @@ public enum MediaUploadError: CustomNSError, LocalizedError {
         switch self {
         case .transferCreationFailed:
             return "transferCreationFailed"
-        case .tranferObservationFailed:
-            return "tranferObservationFailed"
         case .transferFailed(code: _, let description, let isRecoverable):
             return "transferFailed: \(description), recoverable: \(isRecoverable)"
         case .timeout:
@@ -41,8 +37,6 @@ public enum MediaUploadError: CustomNSError, LocalizedError {
         switch self {
         case .transferCreationFailed:
             return "Failed to start the attachment upload"
-        case .tranferObservationFailed:
-            return "Failed to get the status of the attachment upload"
         case .transferFailed(code: _, let description, isRecoverable: _):
             return "Failed to upload the attachment: \(description)"
         case .timeout:
@@ -68,11 +62,11 @@ public class MediaUploader {
 
         return try await withThrowingTaskGroup(of: String.self) { group in
             await withCheckedContinuation { continuation in
-                group.addTask { [weak self] in
-                    guard let self else {
-                        throw CancellationError()
-                    }
-                    return try await self.receivedFinishNotification(for: transferGUID, continuation: continuation)
+                group.addTask {
+                    return try await TransferCenter.receivedFinishNotification(
+                        for: transferGUID,
+                        continuation: continuation
+                    )
                 }
             }
 
@@ -85,12 +79,14 @@ public class MediaUploader {
             group.addTask { [log] in
                 await Task.yield()
                 log.debug("Starting a 30s timeout for the transfer")
-                try await Task.sleep(nanoseconds: 30 * 1000000000)
+                try await Task.sleep(nanoseconds: 30 * 1_000_000_000)
                 log.debug("Reached timeout for the transfer")
                 try Task.checkCancellation()
 
                 log.debug("Checking if the transfer is finished before timing out")
-                if let transfer = transferCenter.transfer(forGUID: transferGUID), let guid = transfer.guid, transfer.isFinished {
+                if let transfer = transferCenter.transfer(forGUID: transferGUID), let guid = transfer.guid,
+                    transfer.isFinished
+                {
                     log.debug("Transfer is finished with state=\(transfer.state) error=\(transfer.error)")
                     return guid
                 }
@@ -103,73 +99,6 @@ public class MediaUploader {
             log.debug("Got finished status from observation")
             return result
         }
-    }
-
-    private func receivedFinishNotification(for transferGUID: String, continuation: CheckedContinuation<Void, Never>) async throws -> String {
-        log.debug("Starting observation task")
-        let updated = NotificationCenter.default.publisher(for: .IMFileTransferUpdated).print("transferUpdated")
-        let finished = NotificationCenter.default.publisher(for: .IMFileTransferFinished).print("transferFinished")
-        let transferEvents = updated.merge(with: finished)
-            .receive(on: DispatchQueue.global(qos: .userInitiated))
-
-        log.debug("Start notification loop")
-        continuation.resume()
-        for try await notification in transferEvents.values {
-            try Task.checkCancellation()
-            log.debug("Handling transfer status event")
-            guard let transfer = notification.object as? IMFileTransfer else {
-                log.error("Got transfer notification with non-transfer data")
-                throw MediaUploadError.tranferObservationFailed
-            }
-
-            guard let guid = transfer.guid else {
-                log.debug("Got transfer notification with nil guid, skipping")
-                continue
-            }
-
-            guard guid == transferGUID else {
-                log.debug("Got notification for guid: \(guid) but interested in \(transferGUID), skipping")
-                continue
-            }
-
-            log.info("Got transfer event notification for: \(guid) with state: \(transfer.state)")
-
-            switch transfer.state {
-            case .finished:
-                log.debug("Transfer \(guid) isFinished: \(transfer.isFinished)")
-                return guid
-            case .error:
-                if transfer.error == 24 {
-                    log.info("Got error 24 when uploading attachment, treating as success")
-                    log.debug(
-                        "Transfer exists at local path: \(transfer.existsAtLocalPath), isFinished: \(transfer.isFinished)"
-                    )
-                    return guid
-                }
-                log.error(
-                    "Transfer \(guid) has an error: \(transfer.error) with description: \(String(describing: transfer.errorDescription))"
-                )
-                throw MediaUploadError.transferFailed(
-                    code: transfer.error,
-                    description: transfer.errorDescription ?? "unknown error",
-                    isRecoverable: false
-                )
-            case .recoverableError:
-                log.error(
-                    "Transfer \(guid) has an recoverable error: \(transfer.error) with description: \(String(describing: transfer.errorDescription))"
-                )
-                throw MediaUploadError.transferFailed(
-                    code: transfer.error,
-                    description: transfer.errorDescription ?? "unknown error",
-                    isRecoverable: true
-                )
-            default:
-                log.debug("Transfer state \(transfer.state) is of no interest to us, skipping")
-                continue
-            }
-        }
-
-        throw MediaUploadError.tranferObservationFailed
     }
 
     @MainActor
